requirements.txt
Django
python-dotenv
privex-helpers
attrs

# Unit testing
django-nose
coverage
codecov
